# perf: optimize symbol library parsing performance by 4x