… Cargo.lock
… kicad_format/src/symbol_library_fast.rs